Problem with type conversions

+++ e2sim/e2sm_examples/kpm_e2sm/src/kpm/bs_connector.hpp
@@ -1,9 +1,9 @@
+#include <chrono>
 #include <thread>
-#include <time>
 
 // send dummy data instead of reading BS metrics
 #define DEBUG 1
 #define LINES_TO_READ 2
 
-void handleTimer(float timer, long ric_req_id);
-void periodicDataReport(float timer, long ric_req_id);
+void handleTimer(int timer, long ric_req_id);
+void periodicDataReport(int timer, long ric_req_id);
